modified onset detection example script
